requirements.txt
# --- Core ---
numpy>=1.24
python-dotenv>=1.0

# --- PostgreSQL + pgvector ---
psycopg2-binary>=2.9
pgvector>=0.2

# --- Embeddings (local) ---
sentence-transformers>=2.6

# --- LangChain ---
langchain>=0.2
langchain-community>=0.2
langchain-core>=0.2 
langchain-groq

# --- Groq API (LLM) ---
groq>=0.9

# --- UI ---
streamlit>=1.32
